Change size of multicast masks macro from 8 to 16 bits.  Update code which used this macro to ensure that it can accept 16 bits worth of multicast masks.  Also update test routines to verify DAR info macros.

+++ fabric_management/librio/inc/RapidIO_Device_Access_Routines_API.h
@@ -202,7 +202,7 @@ uint32_t DAR_add_poreg(DAR_DEV_INFO_t *dev_info, uint32_t oset, uint32_t data);
 uint32_t DAR_get_poreg_idx(DAR_DEV_INFO_t *dev_info, uint32_t oset);
 
 #define NUM_PORTS(x)    (( uint8_t)((((x)->swPortInfo ) & RIO_SW_PORT_INF_TOT) >>  8))
-#define NUM_MC_MASKS(x) ( (uint8_t)((((x)->swMcastInfo) & RIO_SW_MC_INF_MC_MSK)))
+#define NUM_MC_MASKS(x) ( (uint16_t)((((x)->swMcastInfo) & RIO_SW_MC_INF_MC_MSK)))
 #define VEND_CODE(x)    ((uint16_t)(((x)->devID) & RIO_DEV_IDENT_VEND))
 #define DEV_CODE(x)     ((uint16_t)((((x)->devID) & RIO_DEV_IDENT_DEVI) >> 16))
 #define SWITCH(x)	((bool)(((x)->features & RIO_PE_FEAT_SW) ? true:false))
